Moving encryption related properties to CookieCryptTrait
